LayoutTests/fast/mediastream/MediaDevices-getSupportedConstraints-expected.txt

Tests MediaDevices.getSupportedConstraints()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.mediaDevices is an instance of Object
PASS navigator.mediaDevices.getSupportedConstraints is an instance of Function

supportedConstraints = navigator.mediaDevices.getSupportedConstraints()

PASS supportedConstraints.aspectRatio is true
PASS supportedConstraints.deviceId is true
PASS supportedConstraints.echoCancellation is true
PASS supportedConstraints.facingMode is true
PASS supportedConstraints.frameRate is true
PASS supportedConstraints.groupId is true
PASS supportedConstraints.height is true
PASS supportedConstraints.sampleRate is true
PASS supportedConstraints.sampleSize is true
PASS supportedConstraints.volume is true
PASS supportedConstraints.width is true
PASS supportedConstraints.zoom is true

PASS supportedConstraints["aspectRatio"] is true
PASS supportedConstraints["backgroundBlur"] is true
PASS supportedConstraints["deviceId"] is true
PASS supportedConstraints["displaySurface"] is true
PASS supportedConstraints["echoCancellation"] is true
PASS supportedConstraints["facingMode"] is true
PASS supportedConstraints["frameRate"] is true
PASS supportedConstraints["groupId"] is true
PASS supportedConstraints["height"] is true
PASS supportedConstraints["sampleRate"] is true
PASS supportedConstraints["sampleSize"] is true
PASS supportedConstraints["torch"] is true
PASS supportedConstraints["volume"] is true
PASS supportedConstraints["whiteBalanceMode"] is true
PASS supportedConstraints["width"] is true
PASS supportedConstraints["zoom"] is true
PASS successfullyParsed is true

TEST COMPLETE
